V1.1.3
Amplicación campo identificación débito automático

+++ Chevysat Débito Automatico BBVA/DebitoAutomatico.EN/Tablas/ClienteSico.cs
@@ -12,14 +12,13 @@ namespace DebitoAutomatico.EN.Tablas
         public String pOperacion { get; set; }
         public int? pContrato { get; set; }
         public String pTipoDocumento { get; set; }
-        public int? pNroDocumento { get; set; }
+        public Int64? pNroDocumento { get; set; }
         public String pNombreCliente { get; set; }
         public String pEstado { get; set; }
         public String pTipo { get; set; }
         public String pCiudad { get; set; }
         public String pEmail { get; set; }
         public String pCelular { get; set; }
-
         public String pFecha { get; set; }
 
     }
